+++ app/src/main/java/com/techcourse/support/web/filter/ResourceFilter.java
@@ -5,7 +5,6 @@ import jakarta.servlet.annotation.WebFilter;
 import jakarta.servlet.http.HttpServletRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webmvc.org.springframework.web.servlet.mvc.exception.HandlerMappingNotFoundException;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -62,9 +61,6 @@ public class ResourceFilter implements Filter {
     private void doChain(final ServletRequest request, final ServletResponse response, final FilterChain chain) throws IOException, ServletException {
         try {
             chain.doFilter(request, response);
-        } catch (HandlerMappingNotFoundException e) {
-            final RequestDispatcher dispatcher = request.getRequestDispatcher("/404.jsp");
-            dispatcher.forward(request, response);
         } catch (Exception e) {
             log.error("Exception : {}", e.getMessage(), e);
             throw new ServletException(e.getMessage());
